+++ src/main/java/com/example/webnovel/auth/exception/InvalidTokenException.java
@@ -1,4 +1,4 @@
-package com.example.webnovel.global.auth.exception;
+package com.example.webnovel.auth.exception;
 
 import com.example.webnovel.global.error.exception.ErrorCode;
 import com.example.webnovel.global.error.exception.InvalidValueException;
